// functions/portal/src/types.d.ts
interface INotebookBaseDto {
  name: string
  notebookContent: string
}

export interface INotebookUpdateDto extends INotebookBaseDto {
  id: string
  isShared: boolean
}

export interface INotebook extends INotebookBaseDto {
  id: string
  userId: string
}